feat(dashboard): /memories mobile bottom-sheet + coarse-pointer bumps
Adapts /memories to mobile per the shape brief.

  - **MemoryBottomSheet** (mobile detail view, <lg) replaces the
    modal MemoryDetailPanel. Slides up from the viewport bottom,
    80 vh tall, partial backdrop, swipe-handle pill on top edge.
    Built on Radix Dialog primitives so focus trap / Esc /
    backdrop-tap come free; only positioning + slide-in-from-bottom
    differ from the standard editorial Dialog. Reduced-motion
    fallback (animate-in / animate-out utilities honour the media
    query natively).
  - **MemoryDetailContent** extracted so MemoryInspector (rail)
    and MemoryBottomSheet (sheet) wrap the same body. One place
    to evolve the detail-view, two chrome shapes. MemoryInspector
    slimmed to just the rail chrome.
  - **MemoryDetailPanel deleted** — the modal was a transitional
    fallback while waiting for adapt. Sheet replaces it entirely.
  - **Coarse-pointer bumps** wherever a touch target was under 44 px:
    - FilterChips (Active / Add / Overflow): min-h-11, larger
      padding, base text size.
    - Search + Recall inputs: min-h-11, base text size.
    - SortBar selects: h-11, sm text.
    - Bulk-select row checkbox label: min-h-11.
    - Per-row select checkbox: min-w-11 + centered.

Verified at 1440 desktop (Inspector right-rail present) and
iPhone 14 (filter chips wrap to 2 rows, inputs and buttons sized
for thumb). Bottom-sheet visual verify deferred — no seeded data
in dev env to tap. Code path matches the rail (same content,
different chrome).

Typecheck + lint clean; 254 tests pass.
